+++ app/src/main/res/layout/fragment_my_bottom.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.mypage.mine.MyBottomFragment">
+        tools:context=".presentation.ui.mypage.my.MyBottomFragment">
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/tab_profile"
